refactor(portal): extract playlists section into PlaylistsSection.svelte
Moves the read-only "your playlists" section out of the portal page into a
co-located self-loading child component. Parent shrinks 3961 → 3785 lines.

The section owns all the state it touches (playlists, loadingPlaylists) and
its only external dependency is API_URL, so the child self-loads on mount
rather than being driven by the parent's onMount Promise.all. Behavior is
unchanged: loadMyPlaylists already swallowed its own errors via try/catch,
so removing it from the aggregate error path is a no-op.
